docs: ListImgCard 삭제 후 ListColorCard 이름 변경
